TIP-1073: move factory class of the values into write directory
