+++ tests/back/Pim/Enrichment/Specification/Component/Product/Factory/Write/Value/MetricValueFactorySpec.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Specification\Akeneo\Pim\Enrichment\Component\Product\Factory\Value;
+namespace Specification\Akeneo\Pim\Enrichment\Component\Product\Factory\Write\Value;
 
 use Akeneo\Tool\Component\StorageUtils\Exception\InvalidPropertyTypeException;
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Enrichment\Component\Product\Factory\MetricFactory;
-use Akeneo\Pim\Enrichment\Component\Product\Factory\Value\MetricValueFactory;
+use Akeneo\Pim\Enrichment\Component\Product\Factory\Write\Value\MetricValueFactory;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\MetricInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Value\MetricValue;
