+++ tests/back/Pim/Enrichment/Specification/Component/Product/Factory/Write/Value/ScalarValueFactorySpec.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Specification\Akeneo\Pim\Enrichment\Component\Product\Factory\Value;
+namespace Specification\Akeneo\Pim\Enrichment\Component\Product\Factory\Write\Value;
 
 use Akeneo\Tool\Component\StorageUtils\Exception\InvalidPropertyTypeException;
 use PhpSpec\ObjectBehavior;
-use Akeneo\Pim\Enrichment\Component\Product\Factory\Value\ScalarValueFactory;
+use Akeneo\Pim\Enrichment\Component\Product\Factory\Write\Value\ScalarValueFactory;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Value\ScalarValue;
 use Prophecy\Argument;
